requirements.txt
Django==4.1.7
django-environ==0.10.0
django-filter==22.1
djangorestframework==3.14.0
Markdown==3.4.1
Pillow==9.4.0
